Major rework of the Mapping, MappingInput and MappingOutput steps. It needed work all over the place. Unfortunately it's not quite there yet.
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@3899 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/org/pentaho/di/core/Result.java
@@ -49,7 +49,7 @@ public class Result implements Cloneable
 	private Map<String, ResultFile> resultFiles;
 	
 	public boolean stopped;
-    private int nrLinesRejected;
+    private long nrLinesRejected;
 	
 	public Result()
 	{
@@ -389,7 +389,7 @@ public class Result implements Cloneable
     /**
      * @return the nrLinesRejected
      */
-    public int getNrLinesRejected()
+    public long getNrLinesRejected()
     {
         return nrLinesRejected;
     }
@@ -397,7 +397,7 @@ public class Result implements Cloneable
     /**
      * @param nrLinesRejected the nrLinesRejected to set
      */
-    public void setNrLinesRejected(int nrLinesRejected)
+    public void setNrLinesRejected(long nrLinesRejected)
     {
         this.nrLinesRejected = nrLinesRejected;
     }
